+++ semana12/labeddit/src/App.js
@@ -5,9 +5,12 @@ import { BrowserRouter } from 'react-router-dom';
 
 function App() {
   return (
-    <GloblaStateProvider>
-      <Router />
-    </GloblaStateProvider>
+    <BrowserRouter>
+      <GloblaStateProvider>
+        <Header />
+        <Router />
+      </GloblaStateProvider>
+    </BrowserRouter>
   );
 }
