+++ semana12/labeddit/src/Constants/inputs.js
@@ -1,12 +1,12 @@
 export const initialForm = {
     name: '',
-    email:'',
-    password:'',
-    text:'',
-    title:'',
-    comment:'',
-    inputSearch:'',
-    order:'',
+    email: '',
+    password: '',
+    text: '',
+    title: '',
+    comment: '',
+    inputSearch: '',
+    order: '',
 }
 
 export const options = ['Posts recentes', 'Posts antigos', 'Posts mais votados', 'Posts menos votados']
