+++ semana12/labeddit/src/Global/GloblaStateProvider.js
@@ -3,6 +3,8 @@ import { useHistory } from 'react-router';
 import GlobalStateContext from "../Global/GlobalStateContext";
 import { baseUrl, axiosConfig } from '../Constants/api'
 import axios from 'axios';
+import { goToRegister, goToAddPosts, gotToLastPage, goToLogin } from '../Router/coordinator';
+import { Button } from '../Styles/style'
 
 const GlobalStateProvider = (props) => {
     const [posts, setPosts] = useState([])
@@ -10,10 +12,47 @@ const GlobalStateProvider = (props) => {
     const [perPage, setPerPage] = useState(10)
     const [lengthPages, setLengthPages] = useState()
     const [alert, setAlert] = useState(false)
-   
+    const history = useHistory()
+
     useEffect(() => {
         getPosts()
     }, [posts, page, perPage])
+
+    const logout = () => {
+        window.localStorage.removeItem("token");
+        goToLogin(history);
+    };
+
+    const buttonNav = (path) => {
+        switch (path) {
+            case '/':
+                return (
+                    <div>
+                        <Button onClick={() => goToRegister(history)}>Inscreva-se</Button>
+                    </div>
+                )
+            case '/signup':
+                return (
+                    <div>
+                        <Button onClick={() => goToLogin(history)}>Login</Button>
+                    </div>
+                )
+            case '/feed':
+                return (
+                    <div>
+                        <Button onClick={() => goToAddPosts(history)}>Criar Post</Button>
+                        <Button logout onClick={logout}>Logout</Button>
+                    </div>
+                )
+            default:
+                return (
+                    <div>
+                        <Button back onClick={() => gotToLastPage(history)}>Voltar</Button>
+                    </div>
+                )
+        }
+
+    }
 
     const getPosts = async () => {
         try {
@@ -35,14 +74,14 @@ const GlobalStateProvider = (props) => {
             alert(`❌ ${err.response.data.message}`)
         }
     }
-    
+
 
     let start = page * perPage
     let end = start + perPage
 
     const states = { alert, start, end, posts, perPage, page, lengthPages };
-    const setters = { setAlert, setPerPage, setPage, setLengthPages};
-    const requests = { getPosts, votePost };
+    const setters = { setAlert, setPerPage, setPage, setLengthPages };
+    const requests = { getPosts, votePost, buttonNav };
 
     const data = { states, setters, requests };
