+++ semana12/labeddit/src/Hooks/useProtectedPage.js
@@ -1,17 +1,17 @@
-import {useEffect} from 'react'
+import { useEffect } from 'react'
 import { useHistory } from "react-router";
 import { goToLogin } from '../Router/coordinator';
 
 export const useProtectedPage = () => {
-const history = useHistory()
+    const history = useHistory()
 
-useEffect(() => {
-    const token = window.localStorage.getItem('token')
+    useEffect(() => {
+        const token = window.localStorage.getItem('token')
 
-    if(!token) {
-        goToLogin(history)
-    }
-}, [history])
+        if (!token) {
+            goToLogin(history)
+        }
+    }, [history])
 
 }
