+++ semana12/labeddit/src/Pages/Public/ErrorPage.js
@@ -3,15 +3,15 @@ import { Title, Container } from '../../Styles/style';
 import labedditLogo from '../../img/labeddit-logo.png'
 
 function Error() {
-    
+
     return (
         <Container>
-           
-           <img src={labedditLogo} alt='Logo' />
-           <Title error>Ooops! Erro 404</Title>
-           
-           <Title errorPage>Página não encontrada</Title>
-           
+
+            <img src={labedditLogo} alt='Logo' />
+            <Title error>Ooops! Erro 404</Title>
+
+            <Title errorPage>Página não encontrada</Title>
+
         </Container>
     )
 }
