+++ semana12/labeddit/src/Router/Router.js
@@ -1,47 +1,41 @@
-import { BrowserRouter, Route, Switch } from "react-router-dom";
+import { Route, Switch } from "react-router-dom";
 import RegisterPage from '../Pages/Public/RegisterPage'
 import PostsPage from '../Pages/Private/PostsPage'
 import FeedPage from '../Pages/Private/FeedPage'
 import LoginPage from '../Pages/Public/LoginPage'
 import ErrorPage from '../Pages/Public/ErrorPage'
-import Header from '../Components/Header'
 import CommentPage from '../Pages/Private/CommentPage'
 
 function Router() {
     return (
-        <BrowserRouter>
-            <Switch>
-                <Route exact path='/'>
-                    <Header />
-                    <LoginPage />
-                </Route>
 
-                <Route exact path='/signup'>
-                    <Header />
-                    <RegisterPage />
-                </Route>
+        <Switch>
+            <Route exact path='/'>
+                <LoginPage />
+            </Route>
 
-                <Route exact path='/feed'>
-                    <Header />
-                    <FeedPage />
-                </Route>
+            <Route exact path='/signup'>
+                <RegisterPage />
+            </Route>
 
-                <Route exact path='/add/post'>
-                    <Header />
-                    <PostsPage />
-                </Route>
+            <Route exact path='/feed'>
+                <FeedPage />
+            </Route>
 
-                <Route exact path='/post/:postId'>
-                    <Header />
-                    <CommentPage />
-                </Route>
+            <Route exact path='/add/post'>
+                <PostsPage />
+            </Route>
 
-                <Route>
-                    <ErrorPage />
-                </Route>
+            <Route exact path='/post/:postId'>
+                <CommentPage />
+            </Route>
 
-            </Switch>
-        </BrowserRouter>
+            <Route>
+                <ErrorPage />
+            </Route>
+
+        </Switch>
+
     )
 }
